+++ src/mailer/engine.py
@@ -33,6 +33,8 @@ LOCK_WAIT_TIMEOUT = getattr(settings, "MAILER_LOCK_WAIT_TIMEOUT", -1)
 # in the current working directory.
 LOCK_PATH = getattr(settings, "MAILER_LOCK_PATH", None)
 
+logger = logging.getLogger(__name__)
+
 
 def prioritize():
     """
@@ -95,8 +97,8 @@ def _limits_reached(sent, deferred):
     EMAIL_MAX_BATCH = getattr(settings, "MAILER_EMAIL_MAX_BATCH", None)
 
     if EMAIL_MAX_BATCH is not None and sent >= EMAIL_MAX_BATCH:
-        logging.info("EMAIL_MAX_BATCH (%s) reached, "
-                     "stopping for this round", EMAIL_MAX_BATCH)
+        logger.info("EMAIL_MAX_BATCH (%s) reached, "
+                    "stopping for this round", EMAIL_MAX_BATCH)
         return True
 
     # Stop sending emails in the current round if more than X emails get
@@ -104,8 +106,8 @@ def _limits_reached(sent, deferred):
     EMAIL_MAX_DEFERRED = getattr(settings, "MAILER_EMAIL_MAX_DEFERRED", None)
 
     if EMAIL_MAX_DEFERRED is not None and deferred >= EMAIL_MAX_DEFERRED:
-        logging.warning("EMAIL_MAX_DEFERRED (%s) reached, "
-                        "stopping for this round", EMAIL_MAX_DEFERRED)
+        logger.warning("EMAIL_MAX_DEFERRED (%s) reached, "
+                       "stopping for this round", EMAIL_MAX_DEFERRED)
         return True
 
 
@@ -115,13 +117,13 @@ def _throttle_emails():
     EMAIL_THROTTLE = getattr(settings, "MAILER_EMAIL_THROTTLE", 0)
 
     if EMAIL_THROTTLE:
-        logging.debug("Throttling email delivery. "
-                      "Sleeping %s seconds", EMAIL_THROTTLE)
+        logger.debug("Throttling email delivery. "
+                     "Sleeping %s seconds", EMAIL_THROTTLE)
         time.sleep(EMAIL_THROTTLE)
 
 
 def acquire_lock():
-    logging.debug("acquiring lock...")
+    logger.debug("acquiring lock...")
     if LOCK_PATH is not None:
         lock_file_path = LOCK_PATH
     else:
@@ -132,19 +134,19 @@ def acquire_lock():
     try:
         lock.acquire(LOCK_WAIT_TIMEOUT)
     except lockfile.AlreadyLocked:
-        logging.debug("lock already in place. quitting.")
+        logger.error("lock already in place. quitting.")
         return False, lock
     except lockfile.LockTimeout:
-        logging.debug("waiting for the lock timed out. quitting.")
+        logger.error("waiting for the lock timed out. quitting.")
         return False, lock
-    logging.debug("acquired.")
+    logger.debug("acquired.")
     return True, lock
 
 
 def release_lock(lock):
-    logging.debug("releasing lock...")
+    logger.debug("releasing lock...")
     lock.release()
-    logging.debug("released.")
+    logger.debug("released.")
 
 
 def _require_no_backend_loop(mailer_email_backend):
@@ -191,7 +193,7 @@ def send_all():
                 try:
                     if connection is None:
                         connection = get_connection(backend=mailer_email_backend)
-                    logging.info("sending message '{0}' to {1}".format(
+                    logger.info("sending message '{0}' to {1}".format(
                         message.subject,
                         ", ".join(message.to_addresses))
                     )
@@ -207,7 +209,7 @@ def send_all():
                         MessageLog.objects.log(message, RESULT_SUCCESS)
                         sent += 1
                     else:
-                        logging.warning("message discarded due to failure in converting from DB. Added on '%s' with priority '%s'" % (message.when_added, message.priority))  # noqa
+                        logger.warning("message discarded due to failure in converting from DB. Added on '%s' with priority '%s'" % (message.when_added, message.priority))  # noqa
                     message.delete()
 
                 except (socket_error, smtplib.SMTPSenderRefused,
@@ -215,7 +217,7 @@ def send_all():
                         smtplib.SMTPDataError,
                         smtplib.SMTPAuthenticationError) as err:
                     message.defer()
-                    logging.info("message deferred due to failure: %s" % err)
+                    logger.info("message deferred due to failure: %s" % err)
                     MessageLog.objects.log(message, RESULT_FAILURE, log_message=str(err))
                     deferred += 1
                     # Get new connection, it case the connection itself has an error.
@@ -231,9 +233,9 @@ def send_all():
         if use_file_lock:
             release_lock(lock)
 
-    logging.info("")
-    logging.info("%s sent; %s deferred;" % (sent, deferred))
-    logging.info("done in %.2f seconds" % (time.time() - start_time))
+    logger.info("")
+    logger.info("%s sent; %s deferred;" % (sent, deferred))
+    logger.info("done in %.2f seconds" % (time.time() - start_time))
 
 
 def send_loop():
@@ -244,6 +246,6 @@ def send_loop():
 
     while True:
         while not Message.objects.all():
-            logging.debug("sleeping for %s seconds before checking queue again" % EMPTY_QUEUE_SLEEP)
+            logger.debug("sleeping for %s seconds before checking queue again" % EMPTY_QUEUE_SLEEP)
             time.sleep(EMPTY_QUEUE_SLEEP)
         send_all()